feat(76-01): implement BVC classify_volume helper (GREEN phase)
- classify_volume(ohlcv) returns buy_volume + sell_volume columns
- Uses scipy.stats.norm.cdf for Easley/LdP/O'Hara BVC approximation
- Doji bars (high == low) produce 50/50 split via NaN -> fillna(0.5)
- NaN volume bars propagate NaN (no errors)
- All 8 unit tests pass: conservation, bullish/bearish, doji, empty, NaN

+++ tests/test_bvc.py
@@ -83,7 +83,7 @@ def test_volume_conservation(sample_ohlcv: pd.DataFrame):
     result = classify_volume(sample_ohlcv)
     total = result["buy_volume"] + result["sell_volume"]
     pd.testing.assert_series_equal(
-        total, sample_ohlcv["volume"], check_names=False, atol=1e-10
+        total, sample_ohlcv["volume"].astype(float), check_names=False, atol=1e-10
     )
